Finalize TypeScript debugging
