+++ .vscode/cSpell.json
@@ -14,7 +14,8 @@
         "HOMEPATH",
         "USERPROFILE",
         "quickstart",
-        "resizable"
+        "resizable",
+        "Validators"
     ],
     // flagWords - list of words to be always considered incorrect
     // This is useful for offensive words and common spelling errors.
